Airflow/requirements.txt
apache-airflow-providers-google
numpy
pandas
gcsfs
pyarrow
pandas_gbq